+++ HEAANBOOT/src/SecretKey.h
@@ -5,7 +5,6 @@
 
 #include <fstream>
 
-//#include "Functions.h"
 #include "Params.h"
 #include "NumUtils.h"
 
@@ -20,9 +19,9 @@ public:
 
 	SecretKey(Params& params);
 
-	void Write(int SecretKeyID);
+	void Write(long SecretKeyID);
 
-	void Read(int SecretKeyID);
+	void Read(long SecretKeyID);
 };
 
 #endif
